feat redux-saga syntax
