+++ .gitignore
@@ -54,6 +54,7 @@ Thumbs.db
 /node_modules
 /.pnp
 .pnp.js
+.env
 
 # testing
 /coverage
