--- src/store/services/api.jsx
@@ -1,8 +1,0 @@
-import axios from 'axios';
-import configService from './config';
-
-const Api = axios.create({
-  baseUrl: configService.api,
-});
-
-export default Api;
