+++ src/store/services/interceptor.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const api = axios.create({
-  baseURL: 'http://localhost:3000',
+  baseURL: process.env.BASE_URL,
   timeout: 5000,
 });
