+++ src/store/services/testimonials.js
@@ -1,0 +1,8 @@
+import api from './interceptor';
+
+const getAllTestimonials = () => api({
+  method: 'get',
+  url: '/api/testimonials/all',
+});
+
+export default getAllTestimonials;
