+++ src/store/services/users.js
@@ -1,0 +1,8 @@
+import api from './interceptor';
+
+const getUsers = () => api({
+  method: 'get',
+  url: '/api/user',
+});
+
+export default getUsers;
